+++ core/api_container/server/api_container_service.go
@@ -13,6 +13,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/bulk_command_execution_engine"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/bulk_command_execution_engine/v0_bulk_command_execution"
+	"github.com/kurtosis-tech/kurtosis/api_container/server/external_container_store"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/lambda_store"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/lambda_store/lambda_store_types"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network"
@@ -42,6 +43,8 @@ type ApiContainerService struct {
 
 	enclaveDataVolume *enclave_data_volume.EnclaveDataVolume
 
+	externalContainerStore *external_container_store.ExternalContainerStore
+
 	serviceNetwork service_network.ServiceNetwork
 
 	lambdaStore *lambda_store.LambdaStore
@@ -49,9 +52,15 @@ type ApiContainerService struct {
 	bulkCmdExecEngine *bulk_command_execution_engine.BulkCommandExecutionEngine
 }
 
-func NewApiContainerService(enclaveDirectory *enclave_data_volume.EnclaveDataVolume, serviceNetwork service_network.ServiceNetwork, lambdaStore *lambda_store.LambdaStore) (*ApiContainerService, error) {
+func NewApiContainerService(
+	enclaveDirectory *enclave_data_volume.EnclaveDataVolume,
+	externalContainerStore *external_container_store.ExternalContainerStore,
+	serviceNetwork service_network.ServiceNetwork,
+	lambdaStore *lambda_store.LambdaStore,
+) (*ApiContainerService, error) {
 	service := &ApiContainerService{
 		enclaveDataVolume: enclaveDirectory,
+		externalContainerStore: externalContainerStore,
 		serviceNetwork:    serviceNetwork,
 		lambdaStore:       lambdaStore,
 	}
@@ -67,6 +76,27 @@ func NewApiContainerService(enclaveDirectory *enclave_data_volume.EnclaveDataVol
 	service.bulkCmdExecEngine = bulkCmdExecEngine
 
 	return service, nil
+}
+
+func (service ApiContainerService) StartExternalContainerRegistration(ctx context.Context, empty *emptypb.Empty) (*kurtosis_core_rpc_api_bindings.StartExternalContainerRegistrationResponse, error) {
+	registrationKey, ip, err := service.externalContainerStore.StartRegistration()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred starting the registration with the external container store")
+	}
+	result := &kurtosis_core_rpc_api_bindings.StartExternalContainerRegistrationResponse{
+		RegistrationKey: registrationKey,
+		IpAddr:          ip.String(),
+	}
+	return result, nil
+}
+
+func (service ApiContainerService) FinishExternalContainerRegistration(ctx context.Context, args *kurtosis_core_rpc_api_bindings.FinishExternalContainerRegistrationArgs) (*emptypb.Empty, error) {
+	registrationKey := args.RegistrationKey
+	containerId := args.ContainerId
+	if err := service.externalContainerStore.FinishRegistration(registrationKey, containerId); err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred finishing external container registration with registration key '%v' and container ID '%v'", registrationKey, containerId)
+	}
+	return &emptypb.Empty{}, nil
 }
 
 func (service ApiContainerService) LoadLambda(ctx context.Context, args *kurtosis_core_rpc_api_bindings.LoadLambdaArgs) (*emptypb.Empty, error) {