+++ core/cli/commands/root/root.go
@@ -7,6 +7,7 @@ package root
 
 import (
 	"github.com/kurtosis-tech/kurtosis/cli/commands/sandbox"
+	"github.com/kurtosis-tech/kurtosis/cli/commands/test"
 	"github.com/spf13/cobra"
 )
 
@@ -14,8 +15,13 @@ var RootCmd = &cobra.Command{
 	// Leaving out the "use" will auto-use os.Args[0]
 	Use:                        "",
 	Short: "A CLI for interacting with the Kurtosis engine",
+
+	// Cobra will print usage whenever _any_ error occurs, including ones we throw in Kurtosis
+	// This doesn't make sense in 99% of the cases, so just turn them off entirely
+	SilenceUsage: true,
 }
 
 func init() {
 	RootCmd.AddCommand(sandbox.SandboxCmd)
+	RootCmd.AddCommand(test.TestCmd)
 }